% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL10558.R
\docType{data}
\name{GPL10558}
\alias{GPL10558}
\title{Illumina HumanHT-12 v4.0 Expression BeadChip}
\format{
An object of class \code{data.frame} with 35408 rows and 2 columns.
}
\usage{
GPL10558
}
\description{
Illumina HumanHT-12 v4.0 Expression BeadChip
}
\keyword{datasets}
