% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL13497.R
\docType{data}
\name{GPL13497}
\alias{GPL13497}
\title{Agilent-026652 Whole Human Genome Microarray 4x44K v2 (Probe Name Version)}
\format{
An object of class \code{data.frame} with 28573 rows and 2 columns.
}
\usage{
GPL13497
}
\description{
Agilent-026652 Whole Human Genome Microarray 4x44K v2 (Probe Name Version)
}
\keyword{datasets}
