% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL1352.R
\docType{data}
\name{GPL1352}
\alias{GPL1352}
\title{[U133_X3P] Affymetrix Human X3P Array}
\format{
An object of class \code{data.frame} with 43218 rows and 2 columns.
}
\usage{
GPL1352
}
\description{
[U133_X3P] Affymetrix Human X3P Array
}
\keyword{datasets}
