% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL15207.R
\docType{data}
\name{GPL15207}
\alias{GPL15207}
\title{[PrimeView] Affymetrix Human Gene Expression Array}
\format{
An object of class \code{data.frame} with 47097 rows and 2 columns.
}
\usage{
GPL15207
}
\description{
[PrimeView] Affymetrix Human Gene Expression Array
}
\keyword{datasets}
