% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL17077.R
\docType{data}
\name{GPL17077}
\alias{GPL17077}
\title{Agilent-039494 SurePrint G3 Human GE v2 8x60K Microarray 039381 (Probe Name Version)}
\format{
An object of class \code{data.frame} with 31049 rows and 2 columns.
}
\usage{
GPL17077
}
\description{
Agilent-039494 SurePrint G3 Human GE v2 8x60K Microarray 039381 (Probe Name Version)
}
\keyword{datasets}
