% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL1708.R
\docType{data}
\name{GPL1708}
\alias{GPL1708}
\title{Agilent-012391 Whole Human Genome Oligo Microarray G4112A (Feature Number Version)}
\format{
An object of class \code{data.frame} with 27959 rows and 2 columns.
}
\usage{
GPL1708
}
\description{
Agilent-012391 Whole Human Genome Oligo Microarray G4112A (Feature Number Version)
}
\keyword{datasets}
