% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL17586.R
\docType{data}
\name{GPL17586}
\alias{GPL17586}
\title{[HTA-2_0] Affymetrix Human Transcriptome Array 2.0 [Transcript (Gene) Version]}
\format{
An object of class \code{data.frame} with 29201 rows and 2 columns.
}
\usage{
GPL17586
}
\description{
[HTA-2_0] Affymetrix Human Transcriptome Array 2.0 [Transcript (Gene) Version]
}
\keyword{datasets}
