% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL17692.R
\docType{data}
\name{GPL17692}
\alias{GPL17692}
\title{[HuGene-2_1-st] Affymetrix Human Gene 2.1 ST Array [Transcript (Gene) Version]}
\format{
An object of class \code{data.frame} with 27297 rows and 2 columns.
}
\usage{
GPL17692
}
\description{
[HuGene-2_1-st] Affymetrix Human Gene 2.1 ST Array [Transcript (Gene) Version]
}
\keyword{datasets}
