% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL21185.R
\docType{data}
\name{GPL21185}
\alias{GPL21185}
\title{Agilent-072363 SurePrint G3 Human GE v3 8x60K Microarray 039494 [Probe Name Version]}
\format{
An object of class \code{data.frame} with 33581 rows and 2 columns.
}
\usage{
GPL21185
}
\description{
Agilent-072363 SurePrint G3 Human GE v3 8x60K Microarray 039494 [Probe Name Version]
}
\keyword{datasets}
