% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL23126.R
\docType{data}
\name{GPL23126}
\alias{GPL23126}
\title{[Clariom_D_Human] Affymetrix Human Clariom D Assay [Transcript (Gene) Version]}
\format{
An object of class \code{data.frame} with 39899 rows and 2 columns.
}
\usage{
GPL23126
}
\description{
[Clariom_D_Human] Affymetrix Human Clariom D Assay [Transcript (Gene) Version]
}
\keyword{datasets}
