% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL4133.R
\docType{data}
\name{GPL4133}
\alias{GPL4133}
\title{Agilent-014850 Whole Human Genome Microarray 4x44K G4112F (Feature Number Version)}
\format{
An object of class \code{data.frame} with 28956 rows and 2 columns.
}
\usage{
GPL4133
}
\description{
Agilent-014850 Whole Human Genome Microarray 4x44K G4112F (Feature Number Version)
}
\keyword{datasets}
