% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL5175.R
\docType{data}
\name{GPL5175}
\alias{GPL5175}
\title{[HuEx-1_0-st] Affymetrix Human Exon 1.0 ST Array [Transcript (Gene) Version]}
\format{
An object of class \code{data.frame} with 35459 rows and 2 columns.
}
\usage{
GPL5175
}
\description{
[HuEx-1_0-st] Affymetrix Human Exon 1.0 ST Array [Transcript (Gene) Version]
}
\keyword{datasets}
