% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL570.R
\docType{data}
\name{GPL570}
\alias{GPL570}
\title{[HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array}
\format{
An object of class \code{data.frame} with 41492 rows and 2 columns.
}
\usage{
GPL570
}
\description{
[HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array
}
\keyword{datasets}
