% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6102.R
\docType{data}
\name{GPL6102}
\alias{GPL6102}
\title{Illumina Human-6 v2.0 Expression BeadChip}
\format{
An object of class \code{data.frame} with 25589 rows and 2 columns.
}
\usage{
GPL6102
}
\description{
Illumina Human-6 v2.0 Expression BeadChip
}
\keyword{datasets}
