% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6244.R
\docType{data}
\name{GPL6244}
\alias{GPL6244}
\title{[HuGene-1_0-st] Affymetrix Human Gene 1.0 ST Array [Transcript (Gene) Version]}
\format{
An object of class \code{data.frame} with 22511 rows and 2 columns.
}
\usage{
GPL6244
}
\description{
[HuGene-1_0-st] Affymetrix Human Gene 1.0 ST Array [Transcript (Gene) Version]
}
\keyword{datasets}
