% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6254.R
\docType{data}
\name{GPL6254}
\alias{GPL6254}
\title{Phalanx Human OneArray}
\format{
An object of class \code{data.frame} with 21614 rows and 2 columns.
}
\usage{
GPL6254
}
\description{
Phalanx Human OneArray
}
\keyword{datasets}
