% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6480.R
\docType{data}
\name{GPL6480}
\alias{GPL6480}
\title{Agilent-014850 Whole Human Genome Microarray 4x44K G4112F (Probe Name Version)}
\format{
An object of class \code{data.frame} with 30284 rows and 2 columns.
}
\usage{
GPL6480
}
\description{
Agilent-014850 Whole Human Genome Microarray 4x44K G4112F (Probe Name Version)
}
\keyword{datasets}
