% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6848.R
\docType{data}
\name{GPL6848}
\alias{GPL6848}
\title{Agilent-012391 Whole Human Genome Oligo Microarray G4112A (Probe Name Version)}
\format{
An object of class \code{data.frame} with 30284 rows and 2 columns.
}
\usage{
GPL6848
}
\description{
Agilent-012391 Whole Human Genome Oligo Microarray G4112A (Probe Name Version)
}
\keyword{datasets}
