% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6884.R
\docType{data}
\name{GPL6884}
\alias{GPL6884}
\title{Illumina HumanWG-6 v3.0 Expression BeadChip}
\format{
An object of class \code{data.frame} with 30542 rows and 2 columns.
}
\usage{
GPL6884
}
\description{
Illumina HumanWG-6 v3.0 Expression BeadChip
}
\keyword{datasets}
