% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL6947.R
\docType{data}
\name{GPL6947}
\alias{GPL6947}
\title{Illumina HumanHT-12 v3.0 Expression BeadChip}
\format{
An object of class \code{data.frame} with 31006 rows and 2 columns.
}
\usage{
GPL6947
}
\description{
Illumina HumanHT-12 v3.0 Expression BeadChip
}
\keyword{datasets}
