% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-GPL96.R
\docType{data}
\name{GPL96}
\alias{GPL96}
\title{[HG-U133A] Affymetrix Human Genome U133A Array}
\format{
An object of class \code{data.frame} with 19988 rows and 2 columns.
}
\usage{
GPL96
}
\description{
[HG-U133A] Affymetrix Human Genome U133A Array
}
\keyword{datasets}
