% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HG-U133_Plus_2.R
\docType{data}
\name{HG-U133_Plus_2}
\alias{HG-U133_Plus_2}
\title{[HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array}
\format{
An object of class \code{environment} of length 54675.
}
\usage{
`HG-U133_Plus_2`
}
\description{
[HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array
}
\keyword{datasets}
