% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HG-U133A.R
\docType{data}
\name{HG-U133A}
\alias{HG-U133A}
\title{[HG-U133A] Affymetrix Human Genome U133A Array}
\format{
An object of class \code{environment} of length 22283.
}
\usage{
`HG-U133A`
}
\description{
[HG-U133A] Affymetrix Human Genome U133A Array
}
\keyword{datasets}
