% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-PrimeView.R
\docType{data}
\name{PrimeView}
\alias{PrimeView}
\title{[PrimeView] Affymetrix Human Gene Expression Array}
\format{
An object of class \code{environment} of length 49395.
}
\usage{
PrimeView
}
\description{
[PrimeView] Affymetrix Human Gene Expression Array
}
\keyword{datasets}
