% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-U133_X3P.R
\docType{data}
\name{U133_X3P}
\alias{U133_X3P}
\title{[U133_X3P] Affymetrix Human X3P Array}
\format{
An object of class \code{environment} of length 61359.
}
\usage{
U133_X3P
}
\description{
[U133_X3P] Affymetrix Human X3P Array
}
\keyword{datasets}
